Make sesolve and mesolve independent from odeint
